Wrap the unlink in a try...catch
Right now the module freaks when unlinking if the file does not exist.
That does not sound like a good idea since the main point is to delete
the "fail fast" file and there is nothing to delete so the goal has
already been acomplished. Additionally, freaking out just creates a
messy log for the users and that is a bad time for everyone when
later working with troubleshooting actual issues. (You do not want
random Errors from some modules you are using when your real build is
going haywire!)